+++ runs/scraper_scripts/compile_database.py
@@ -1,7 +1,5 @@
-import zap_scrap as zs
-
 category = 'oxfords'
 save_dir = 'databases_new'
-zs.dataIO.compile_database(save_dir, category)
+shoezam.zap_scrap.dataIO.compile_database(save_dir, category)
