+++ runs/scraper_scripts/test_ImageFixer.py
@@ -1,6 +1,3 @@
-import zap_scrap as zs
-
-
-data_dir = './databases/oxfords_db.json' 
+data_dir = './databases/oxfords_db.json'
 image_dir = './images/'
-crawler = zs.ImageFixer.ImageFixer(data_dir, image_dir)
+crawler = shoezam.zap_scrap.ImageFixer.ImageFixer(data_dir, image_dir)
