+++ runs/scraper_scripts/test_ParseProduct.py
@@ -1,5 +1,5 @@
 #from zap_scrap import ParseProduct.ParseProduct
-import zap_scrap as zs
+from shoezam import zap_scrap as zs
 
 product = zs.ParseProduct('https://www.zappos.com/p/cole-haan-original-grand-shortwing-woodbury-leather-java/product/8931778/color/646844')
 product.print_product_details()
