+++ shoezam/shoe_classifier/add_to_path.sh
